fix: fixed switching between text submission to call overlay issue
